Fixes relative imports
Removes forward slash from the paths being retrived from `opts.from` array.
The file resolving was not successful with the extra slash.

I'm not sure this is the correct fix. I'm not even sure this is an actual bug.
It seems to be a solution to my problem though:
"I want to @import files with a relative path, but the file being imported was
not being found because the path was not correct (it had the extra forward slash
in the front of the path)"